Rename: repository -> domain 폴더명 변경
