+++ src/main/java/roomescape/domain/Reservation.java
@@ -1,4 +1,4 @@
-package roomescape.repository;
+package roomescape.domain;
 
 public record Reservation(Long id, String name, String date, String time) {
